Consolidate emergent mind utilities
