+++ pipelines/__init__.py
@@ -3,9 +3,9 @@
 The modules provide simple building blocks for the meta-autobiography engine:
 ``maintenance_tools`` now consolidates repository distillation, history
 extraction, token summaries and graph building alongside diff utilities.
-``memory_graph_builder`` links concepts, ``affective_oracle`` infers mood,
-``braided_mind_dueler`` offers parallel reasoning, ``quantum_dreamweaver``
-sparks creativity, and ``collapse_repo`` prunes the repository when ready.
+``memory_graph_builder`` links concepts while ``emergent_mind`` now bundles
+emotion inference, braided reasoning, spontaneous dreaming and the full
+orchestrator cycle. ``collapse_repo`` prunes the repository when ready.
 
 ``EXCLUDE_PATHS`` enumerates files or folders that should remain untouched when
 distilling or collapsing the repository. Modules importing this package can rely
